% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_sp_linking.R
\name{chipSPLink}
\alias{chipSPLink}
\title{ChIP-SP Core Spatial Integration and Ranking}
\usage{
chipSPLink(chip_file, hic_df)
}
\arguments{
\item{chip_file}{Character. ChIP-seq peak file (*ChIP.xls).}

\item{hic_df}{data.frame. Output from mergeHiCLoops().}
}
\value{
A data.frame of spatially linked regions ranked by ChIP-SP score.
}
\description{
Integrates ChIP-seq peaks with Hi-C chromatin loops by assigning
spatially linked regulatory regions and ranking interactions
based on ChIP pileup and Hi-C loop confidence.
}
\examples{
hic_df <- mergeHiCLoops(list.files(pattern = "HiC\\\\.xls$", full.names = TRUE))
res <- chipSPLink("TF_ChIP.xls", hic_df)

}
